changed latitude and longitude data types to decimal as found in some sources

+++ database/migrations/2024_03_24_045148_create_events_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             $table->string('state');
             $table->string('postalCode');
             $table->string('country');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->decimal('latitude', 10, 7);
+            $table->decimal('longitude', 10, 7);
             $table->text('description');
             $table->text('speakers');
             $table->timestamps();
